bugfix: add delay between teensy col write and pin read
without this we have ghosting problems on the bottommost keys of the
teensy side

thanks to hasu (on geekhack) for the suggestion, and PrinsValium for
confirming erratic behavior with his firmware without these delays.
thanks DOX for making the changes and trying it out.  i'm just adding it
to the repo.