layout-gen: still testing stuff
will probably have to take a break for a while

javascript looks promising :) .  this application just lends itself so
well to state!  or maybe it's because i'm not used to programming in a
pure functional language like Elm...  Elm is fun though :)

decided not to write a second test for Elm.  partly because i didn't
have time.  what i wanted to try was doing things based on the Elm
calculator example.  I haven't tried using buttons yet, and they might
work out.  then again, it might turn out to be a whole lot easier in
javascript.  and then it would be a bit more portable as well...

so i suppose i'm leaning towards javascript.  we'll see :)