New TYPE__ macros for typing the key directly on key press
I needed to introduce these, because I had a problem with typing "<>".

On my PC configured to use German keyboard layout following happened:

The key ">" was not processed by the OS when "<" was typed just before.
The reason I think is, that normally on a German keyboard "<" and ">"
are on the same key (the key right of the left "shift").

For "<" you press that key. For ">" you have to first release the key
press and hold "shift", and press the key again. This was not possible
with the KEYS__DEFAULT and KEYS__SHIFTED macros.

Therefore I introduced the TYPE__ macros. They do all the work, i.e.
send a press followed by a release to the OS. With these macros it is
not possible any more to use repeating of keys at the OS level, but for
me this is no problem, because I don't use this for normal keys. Maybe
this is a problem with shortcut combinations, where multiple keys need
to be pressed simultaniously.

Since the ALT_GR macros where also doing all the work on key press, I
renamed them to TYPE__, to reflect this.